Got self movement working.
Took the move code out of the keyPressed event and moved into a method that could be called from within the SnakePanel
Got the timer working. Eclipse strugglied with the auto includes on that one.
Got the gitignore working for the bin paths. Weird that it was able to understand linux slashes.